Update and rename WebSocketTradeUnboundedSource.kt to ExchangeClientUnboundedSource.kt
